final solution with main solution.rb file

+++ app/file_helper.rb
@@ -1,7 +1,12 @@
 module File_helper
 
   def load_file(path,rules)
+
     result = {}
+    if !File.exist?(path)
+      puts "Wrong input.Source file does not exist"
+      exit 0
+    end
     File.open(path, "r").each_line do |line|
       for title,pattern in rules do 
         result[title] ||= []
@@ -15,8 +20,13 @@ module File_helper
     result
   end
   
-
-  def save_file(path)
+  def save_file(path,array)
+    File.open(path,'w') do |f|
+      array.each do |el| 
+        f.write(el)
+        f.write(' ')
+      end
+    end
   end
